# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DYL Real-Estate Services API
  description: >
    Admin-verified real estate listings API (Lagos-first, Nigeria). Three
    roles share this API: **seekers** (browse/save/inquire, the default
    `role` on every account), **owners** (list/manage their own
    properties — a seeker becomes an owner the moment they create a
    listing), and **admins** (`role: admin`) who review and verify
    listings and moderate the platform.

    Authentication is a bearer JWT returned by `/auth/login`,
    `/auth/register`, or `/auth/google`. Send it as
    `Authorization: Bearer <token>` on every authenticated request.
    Tokens are invalidated (not just expired) when the account's
    password changes — see `token_version` under Auth below.
  version: "1.0.0"
servers:
  - url: /api
    description: Relative to wherever this API is deployed (see README for
      environment-specific base URLs).
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Obtained from `/auth/login`, `/auth/register`, or `/auth/google`.
        Carries `id`, `role`, `email`, `name`, and `token_version` — the
        last of which is checked against the user's current value on
        every request, so it's what lets a password change or reset
        immediately invalidate every previously-issued token rather than
        waiting out the token's natural expiry.
...
  responses:
    BadRequest:
      description: Validation failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, expired, or revoked token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    TooManyRequests:
      description: Rate limited. Sensitive auth endpoints (login, register,
        password reset, etc) and public write endpoints (inquiries,
        inspection requests) are IP-rate-limited on top of any
        account-level lockout.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Locked:
      description: Account temporarily locked after repeated failed login
        attempts (independent of, and in addition to, the IP-based rate
        limit above). Try again after the cooldown named in the error
        message.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: "Invalid email or password." }
...
    User:
      type: object
      description: Never includes `password_hash` — every endpoint that
        returns a user strips it before responding.
      properties:
        id: { type: integer }
        name: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        role: { type: string, enum: [user, admin] }
        user_type: { type: string, enum: [user, landlord, agent, developer] }
        email_verified: { type: boolean }
        email_opt_in: { type: boolean }
        created_at: { type: string, format: date-time }
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Create an account
      description: Rate-limited (IP-based). Sends a verification email
        (best-effort — a delivery failure doesn't fail registration
        itself; see `/auth/resend-verification`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [first_name, last_name, email, password]
              properties:
                first_name: { type: string }
                last_name: { type: string }
                email: { type: string, format: email }
                password:
                  type: string
                  minLength: 8
                  maxLength: 72
                  description: >
                    Also rejected: well-known common passwords, and
                    anything containing the account's own name or email
                    local-part.
                user_type: { type: string, enum: [user, landlord, agent, developer] }
                phone: { type: string }
                email_opt_in: { type: boolean }
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: Email already registered.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in
      description: >
        IP rate-limited AND account-level-locked: 5 failed attempts on
        the *same account* locks it for 15 minutes regardless of which
        IPs the attempts came from — a second, independent layer on top
        of the IP limiter, since the IP limiter alone doesn't stop a
        distributed attempt against one specific account.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Logged in.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Wrong email or password. Deliberately identical
            whether the email doesn't exist or the password is wrong —
            and never reveals how many attempts remain, since that would
            help an attacker time around the lockout.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "423": { $ref: "#/components/responses/Locked" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/me:
    get:
      tags: [Auth]
      summary: Get the current user
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change password (while logged in)
      description: Bumps `token_version`, invalidating every other token
        issued before this point (e.g. a session left open on another
        device) — then immediately re-signs a fresh token for *this*
        request so the caller isn't logged out of their own session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password:
                  type: string
                  minLength: 8
                  maxLength: 72
                  description: Same policy as registration - also rejected
                    if it's a well-known common password or contains the
                    account's own name or email.
      responses:
        "200":
          description: Changed. Returns a new token — the old one is now
            rejected by every other endpoint.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  token: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Not authenticated, or `current_password` is wrong.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset email
      description: Always returns the same generic message whether or not
        the email exists, so this endpoint can't be used to enumerate
        registered accounts.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Generic confirmation (see description above).
          content:
            application/json:
              schema:
                type: object
                properties: { message: { type: string } }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Complete a password reset
      description: Also bumps `token_version`, so any token issued before
        the reset stops working — the account owner has to log in fresh.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token: { type: string, description: From the emailed reset link. }
                new_password:
                  type: string
                  minLength: 8
                  maxLength: 72
                  description: Same policy as registration - also rejected
                    if it's a well-known common password or contains the
                    account's own name or email.
      responses:
        "200":
          description: Password reset.
          content:
            application/json:
              schema:
                type: object
                properties: { message: { type: string } }
        "400":
          description: Token is invalid, already used, or expired (1 hour).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Confirm an email address
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, description: From the emailed verification link. }
      responses:
        "200":
          description: Verified.
          content:
            application/json:
              schema:
                type: object
                properties: { message: { type: string } }
        "400":
          description: Token is invalid, already used, or expired (24 hours).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend the verification email
      responses:
        "200":
          description: Sent (or already verified — same 200, message differs).
          content:
            application/json:
              schema:
                type: object
                properties: { message: { type: string } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/google:
    post:
      tags: [Auth]
      summary: Sign in (or register) with Google
      description: Verifies the ID token directly against Google rather
        than doing JWT signature verification locally. An account
        created or linked this way is auto-marked `email_verified`,
        since Google already confirmed the address.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credential]
              properties:
                credential: { type: string, description: The Google ID token. }
      responses:
        "200":
          description: Signed in / account created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: Google couldn't verify the token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503":
          description: Google sign-in isn't configured on this deployment
            (`GOOGLE_CLIENT_ID` unset).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...